requirement.txt
functions-framework==3.*
google-cloud-firestore
requests